requirements.txt
python-bitvavo-api
python-dotenv
